Add tinythreads module and test prog for it

+++ lib/embed/timer.h
@@ -12,9 +12,9 @@ typedef enum counter {TIMER1, WDT} counter_t;
 
 uint8_t tick;
 
-void init_timer(uint8_t rate, counter_t type);
-void init_wdt(uint8_t rate);
-void init_timer1(uint8_t rate);
+void init_timer(uint16_t rate, counter_t type);
+void init_wdt(uint16_t rate);
+void init_timer1(uint16_t rate);
 
 #endif //_TIMER_H_
